==> script/6a_DEG_epi.ipynb <==
suppressMessages(library('tidyverse'))
#library('ggplot2')
#library('dplyr')
suppressMessages(library('patchwork'))
suppressMessages(library('cowplot'))

library('rhdf5')

suppressMessages(library('Seurat'))
suppressMessages(library('DESeq2'))

library('readxl')
library('scuttle')
library('scran')
library('Matrix.utils')
library('tmod')
library('ggrepel')
library("GGally")
library("ggpubr")

library('msigdbr')
library('org.Hs.eg.db')
library('GO.db')

sessionInfo()

suppressMessages(library(scales))

options(repr.plot.width=12, 
        repr.plot.height=8,
        repr.plot.res = 200)

plot_theme = theme_bw() + 
             theme(title = element_text(size = 30, hjust = 0.5),
                   axis.title = element_text(size = 28),
                   text = element_text(size=26),
                   plot.margin = margin(0,0,0,0,"cm")) 

plot_theme_tmod = theme_bw() + 
                  theme(title = element_text(size = 14),
                   axis.title = element_text(size = 14),
                   axis.text = element_text(size = 12),
                   legend.title =  element_text(size = 12),
                   text = element_text(size = 10),
                   plot.margin = margin(0,0,0,0,"cm")) 

'%notin%' = Negate('%in%')

source('function_single_cell_conversion.R')

# 120
R_list = c(#'WNT1','WNT2','WNT2B','WNT3','WNT3A','WNT4',
            #              'WNT5A','WNT5B','WNT6','WNT7A','WNT7B',#'WNT8A',
            #              'WNT8B','WNT9A','WNT9B','WNT10A','WNT10B','WNT11','WNT16', 
            #'RSPO1','RSPO2','RSPO3','RSPO4',
           'FZD1','FZD2','FZD3','FZD4','FZD5','FZD6','FZD7','FZD8','FZD9','FZD10',
                        'LGR4','LGR5','LGR6',
           #'EGF','TGFA','HBEGF','BTC','AREG','EREG','EPGN',
           'EGFR','ERBB2','ERBB3','ERBB4',
           #'FGF1','FGF2','FGF3','FGF4','FGF5',#'FGF6',
           #   'FGF7','FGF8','FGF9','FGF10',
           'FGFR1','FGFR2','FGFR3','FGFR4',
           #'BMP2','BMP4','BMP6','BMP7', 
           'BMPR1A','BMPR1B','BMPR2',
           #'TGFB1','TGFB2','TGFB3','LTBP1','LRRC32','BMP3',#'GDF2','BMP10',
           #    'INHA','INHBA','INHBB',
           #    'INHBC','INHBE','FST',
           'TGFBR1','TGFBR2','ACVR1','ACVR1B','ACVR1C',#'AVCR2A',
                 'ACVR2B',
           #'JAG1','JAG2','DLL1','DLL3','DLL4',
           'NOTCH1','NOTCH2','NOTCH3','NOTCH4',
           #'EFNA1','EFNA2','EFNA3','EFNA4','EFNA5','EFNB1','EFNB2','EFNB3',
           'EPHA1','EPHA2','EPHA3','EPHA4',
                'EPHA5','EPHA6','EPHA7','EPHA8',
                'EPHB1','EPHB2','EPHB3','EPHB4',#'EPHB5',
                'EPHB6',
           'DCHS1', 'DCHS2',
           'FAT1', 'FAT2', 'FAT3', 'FAT4'
           )

new_data_folder = '../processed_data'

h5_path = '../processed_data/CB_epi_Numbat_Scitcem_inferCNV_icms_Uhlitz_resolved_identity.h5'

# h5
adata_h5 = H5Fopen(h5_path)

# main matrix
adata_all = read_scanpy_h5(h5_path,
                           if_pca = TRUE, if_umap = TRUE, if_CB_counts=TRUE, if_raw_counts=FALSE)

adata_all

# get sample label
cat = c(adata_h5$obs$'sample'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'sample'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'sample')

# get patient 
cat = c(adata_h5$obs$'patient'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'patient'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'patient')

# get sample_origin 
cat = c(adata_h5$obs$'sample_origin'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'sample_origin'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'sample_origin')

# get MS_status
cat = c(adata_h5$obs$'MS_status'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'MS_status'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'MS_status')

# get cna_clone
cat = c(adata_h5$obs$'inferCNV_result'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'inferCNV_result'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'inferCNV_result')

# get scitcem_call
cat = c(adata_h5$obs$'scitcem_call'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'scitcem_call'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'scitcem_call')

# get numbat_com
cat = c(adata_h5$obs$'numbat'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'numbat'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'numbat')

# get scANVI_icms
cat = c(adata_h5$obs$'iCMS_scANVI'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'iCMS_scANVI'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'iCMS_scANVI')

# get Uhlitz_scANVI
cat = c(adata_h5$obs$'Uhlitz_scANVI'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'Uhlitz_scANVI'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'Uhlitz_scANVI')

# get cell_identity
cat = c(adata_h5$obs$'cell_identity'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'cell_identity'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'cell_identity')

# get louvain
cat = c(adata_h5$obs$'louvain'$categories)
cat_obs = data.frame(cat_name = cat[apply(adata_h5$obs$'louvain'$codes, 1, function(x)x+1)],
                                              row.names = adata_h5$obs$`_index`)

adata_all = AddMetaData(object = adata_all, metadata = cat_obs, col.name = 'louvain')

h5closeAll()

# cell type in Uhlitz
cell_type_list = c(unique(adata_all@meta.data['Uhlitz_scANVI'])$Uhlitz_scANVI) %>% sort()
cell_type_list = gsub('/', '_', cell_type_list)
cell_type_list = gsub(' ', '_', cell_type_list)
cell_type_list

# replace the / cuz it will cuz a file saving problem
adata_all@meta.data  = adata_all@meta.data %>% mutate(Uhlitz_scANVI = gsub('/', '_', Uhlitz_scANVI)) %>%
mutate(Uhlitz_scANVI = gsub(' ', '_', Uhlitz_scANVI))

table(adata_all@meta.data['cell_identity'])

adata_all@meta.data = adata_all@meta.data %>% 
mutate('GT_other' = case_when(cell_identity == 'genomically_tumour' ~ 'genomically_tumour',
                                        .default = 'others'))

cell_type_order = c('Tumour-like',
                    'Stem', 'Stem_TA', 
                    'Tuft',
                    'Immature_Goblet', 'Goblet',
                    'Enterocyte_progenitor', 'Enterocytes')

### remove TCs from normal sample and genomically normal

adata_GT = subset(adata_all, subset = sample_origin == 'tumour\nsample' & 
                      cell_identity == 'genomically_tumour')

adata_GN_noTCs = subset(adata_all, subset = sample_origin == 'tumour\nsample' & 
                        cell_identity == 'genomically_normal' &
                        Uhlitz_scANVI %notin% c('TC1', 'TC2', 'TC3', 'TC4'))

adata_n_noTCs = subset(adata_all, subset = sample_origin != 'tumour\nsample'  & 
                       Uhlitz_scANVI %notin% c('TC1', 'TC2', 'TC3', 'TC4'))

adata_GT@meta.data = adata_GT@meta.data %>% mutate(Uhlitz_scANVI = 'Tumour-like')

adata_GN_noTCs

adata_n_noTCs

adata = merge(adata_GT,adata_GN_noTCs)

adata = merge(adata,adata_n_noTCs)

### find markers for louvain clusters

adata = SetIdent(adata, value = 'louvain')

specific_to_9 = FindMarkers(adata, ident.1 = '9')

specific_to_9

options(repr.plot.width=5, 
        repr.plot.height=5,
        repr.plot.res = 300)

specific_to_9 %>%  
tibble::rownames_to_column("gene_symbol") %>%
ggplot(aes(x = avg_log2FC, y = -log(p_val_adj),
               color = I(ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, 'navyblue', 'grey')),
               label = ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, gene_symbol,''))) +
geom_point(size = 1) + 
plot_theme_tmod  + 
geom_text_repel(force = 2, force_pull = 0.5, 
                box.padding = 1, point.padding = 1,
                max.overlaps = 20, direction = 'both',
                min.segment.length = 0.1,
                aes(color = 'grey10', segment.linetype = 3)) +
labs(title = paste0('Volcano plot of cluster 9 vs others')) +
theme(plot.title = element_text(hjust = 0.5)) +
xlim(-2,2)
    

options(repr.plot.width=16, 
        repr.plot.height=8,
        repr.plot.res = 300)

adata = SetIdent(adata, value = 'louvain')
DotPlot(adata, features = c(rownames_to_column(specific_to_9)[1:20, 'rowname'])) +
theme(axis.text.x = element_text(angle = 90))

# dot plot of the top genes across all clusters
options(repr.plot.width=8, 
        repr.plot.height=6,
        repr.plot.res = 300)

c9_marker_exp = as.data.frame(FetchData(adata, vars = rownames_to_column(specific_to_9)[1:20, 'rowname'])) %>% 
rownames_to_column()

df =  as.data.frame(adata@meta.data[c('louvain')])  %>% rownames_to_column()

#df$cell_identity = factor(cell_type_df$cell_identity, 
                                    #levels = c('normal_sample', 'genomically_normal', 'genomically_tumour')) 

#df$Uhlitz_scANVI = factor(cell_type_df$Uhlitz_scANVI, 
                                         #levels = cell_type_order)
df$louvain = factor(df$louvain, levels = c('0','1','2','3','4','5','6','7','8','9',
                                           '10','11','12','13','14','15','16','17','18','19'))



c9_marker_exp %>% 
left_join(df, by = c('rowname' = 'rowname')) %>%
pivot_longer(!c(rowname, louvain), 
             names_to = 'gene_symbol', values_to = 'expression') %>%
mutate(gene_symbol = factor(gene_symbol, levels = rownames_to_column(specific_to_9)[1:20, 'rowname'])) %>%
group_by(gene_symbol, louvain) %>%
mutate(mean_expression = mean(expression), 
       ncell = n(), 
       nonzero = ifelse(expression != 0, 1, 0),
       nonzerocell = sum(nonzero),
       cell_fraction = nonzerocell/ncell) %>%
ungroup() %>%
distinct() %>%
ggplot(aes(x = louvain, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = cell_fraction, color = mean_expression)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90),
      strip.background = element_blank(), 
      strip.text = element_text(size = 13), 
      panel.spacing = unit(1, "lines")) +
geom_vline(xintercept = seq(0.5, 20), color = 'grey') +
geom_hline(yintercept = seq(0.5, 54), color = 'grey') +
scale_y_discrete(limits = rev)

# dot plot of the top genes across all clusters
options(repr.plot.width=8, 
        repr.plot.height=6,
        repr.plot.res = 300)

sig_c9_list= specific_to_9 %>% filter(abs(avg_log2FC) > 0.5) %>% rownames_to_column()

c9_marker_exp = as.data.frame(FetchData(adata, vars = sig_c9_list$rowname)) %>% 
rownames_to_column()

df =  as.data.frame(adata@meta.data[c('louvain')])  %>% rownames_to_column()

#df$cell_identity = factor(cell_type_df$cell_identity, 
                                    #levels = c('normal_sample', 'genomically_normal', 'genomically_tumour')) 

#df$Uhlitz_scANVI = factor(cell_type_df$Uhlitz_scANVI, 
                                         #levels = cell_type_order)
df$louvain = factor(df$louvain, levels = c('0','1','2','3','4','5','6','7','8','9',
                                           '10','11','12','13','14','15','16','17','18','19'))



c9_marker_exp %>% 
left_join(df, by = c('rowname' = 'rowname')) %>%
pivot_longer(!c(rowname, louvain), 
             names_to = 'gene_symbol', values_to = 'expression') %>%
mutate(gene_symbol = factor(gene_symbol, levels = sig_c9_list$rowname)) %>%
group_by(gene_symbol, louvain) %>%
mutate(mean_expression = mean(expression), ncell = n(), 
       nonzero = ifelse(expression != 0, 1, 0),
       nonzerocell = sum(nonzero),
       cell_fraction = nonzerocell/ncell) %>%
ungroup() %>%
distinct() %>%
ggplot(aes(x = louvain, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = cell_fraction, color = mean_expression)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90),
      strip.background = element_blank(), 
      strip.text = element_text(size = 13), 
      panel.spacing = unit(1, "lines")) +
geom_vline(xintercept = seq(0.5, 20), color = 'grey') +
geom_hline(yintercept = seq(0.5, 54), color = 'grey') +
#facet_grid(cols = vars(cell_identity), space = 'free', scale = 'free') +
scale_y_discrete(limits = rev)

options(repr.plot.width=24, 
        repr.plot.height=6,
        repr.plot.res = 300)

stem_markers = c('LGR5', 'ASCL2', 'EPHB2', 'OLFM4', 'SOX9', 'TNFRSF19', 'BCL2', 'SOX4', 'CD44',
                'YAP1', 'KIT', 'JUN','JUNB', 'MIF', 'NPM1')

stem_marker_exp = as.data.frame(FetchData(adata, vars = stem_markers)) %>% 
rownames_to_column()

df =  as.data.frame(adata@meta.data[c('louvain', 'cell_identity')])  %>% rownames_to_column()

df$louvain = factor(df$louvain, levels = c('0','1','2','3','4','5','6','7','8','9',
                                           '10','11','12','13','14','15','16','17','18','19'))

df$cell_identity = factor(df$cell_identity, 
                                    levels = c('normal_sample', 'genomically_normal', 'genomically_tumour')) 
#cell_type_df$Uhlitz_scANVI = factor(cell_type_df$Uhlitz_scANVI, 
#                                         levels = cell_type_order)


stem_marker_exp %>% 
left_join(df, by = c('rowname' = 'rowname')) %>%
pivot_longer(!c(rowname, louvain, cell_identity), 
             names_to = 'gene_symbol', values_to = 'expression') %>%
mutate(gene_symbol = factor(gene_symbol, levels = stem_markers)) %>%
group_by(gene_symbol, louvain,cell_identity) %>%
mutate(mean_expression = mean(expression), ncell = n(), 
       nonzero = ifelse(expression != 0, 1, 0),
       nonzerocell = sum(nonzero),
       cell_fraction = nonzerocell/ncell) %>%
ungroup() %>%
distinct() %>%
ggplot(aes(x = louvain, y = gene_symbol)) +
geom_tile(fill = 'white') +
#coord_equal() +
geom_point(aes(size = cell_fraction, color = mean_expression)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90),
      strip.background = element_blank(), 
      strip.text = element_text(size = 13), 
      panel.spacing = unit(1, "lines")) +
geom_vline(xintercept = seq(0.5, 20), color = 'grey') +
geom_hline(yintercept = seq(0.5, 54), color = 'grey') +
scale_y_discrete(limits = rev) +
facet_grid(cols = vars(cell_identity))

# dot plot of the top genes across all clusters
options(repr.plot.width=48, 
        repr.plot.height=16,
        repr.plot.res = 300)

stem_markers = c('LGR5', 'ASCL2', 'EPHB2', 'OLFM4', 'SOX9', 'TNFRSF19', 'BCL2', 'SOX4', 'CD44',
                'YAP1', 'KIT', 'JUN','JUNB', 'MIF', 'NPM1')

stem_marker_exp = as.data.frame(FetchData(adata, vars = stem_markers)) %>% 
rownames_to_column()

df =  as.data.frame(adata@meta.data[c('louvain', 'cell_identity','Uhlitz_scANVI')])  %>% rownames_to_column()

df$louvain = factor(df$louvain, levels = c('0','1','2','3','4','5','6','7','8','9',
                                           '10','11','12','13','14','15','16','17','18','19'))

df$cell_identity = factor(df$cell_identity, 
                                    levels = c('normal_sample', 'genomically_normal', 'genomically_tumour')) 
df$Uhlitz_scANVI = factor(df$Uhlitz_scANVI, 
                                         levels = cell_type_order)


stem_marker_exp %>% 
left_join(df, by = c('rowname' = 'rowname')) %>%
pivot_longer(!c(rowname, louvain, cell_identity, Uhlitz_scANVI), 
             names_to = 'gene_symbol', values_to = 'expression') %>%
mutate(gene_symbol = factor(gene_symbol, levels = stem_markers)) %>%
group_by(gene_symbol, louvain, cell_identity, Uhlitz_scANVI) %>%
mutate(mean_expression = mean(expression), ncell = n(), 
       nonzero = ifelse(expression != 0, 1, 0),
       nonzerocell = sum(nonzero),
       cell_fraction = nonzerocell/ncell) %>%
ungroup() %>%
distinct() %>%
ggplot(aes(x = louvain, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = cell_fraction, color = mean_expression)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90),
      strip.background = element_blank(), 
      strip.text = element_text(size = 13), 
      panel.spacing = unit(1, "lines")) +
geom_vline(xintercept = seq(0.5, 20), color = 'grey') +
geom_hline(yintercept = seq(0.5, 54), color = 'grey') +
scale_y_discrete(limits = rev) +
facet_grid(rows = vars(cell_identity), cols = vars(Uhlitz_scANVI))

between_7_and_9 = FindMarkers(adata, ident.1 = '9', ident.2 = '7')

between_7_and_9

# take all cells in cluster 9, and find markers that separate cells in the ident group
T_vs_N_in_9 = FindMarkers(adata, subset.ident = '9', group.by = 'sample_origin', ident.1 = 'tumour\nsample')

T_vs_N_in_9

Not enough cells
T_vs_N_in_9_in_stems = FindMarkers(subset(adata_all, Uhlitz_scANVI %in% c('Stem', 'Stem_TA')), 
                          subset.ident = '9', group.by = 'sample_origin', ident.1 = 'tumour\nsample')

adata = SetIdent(adata, value = 'Uhlitz_scANVI')

# normal stem vs tumour stem
T_vs_N_stems = FindMarkers(adata,
                           subset.ident = c('Stem'), 
                           group.by = 'sample_origin', ident.1 = 'tumour\nsample')

T_vs_N_stems

options(repr.plot.width=6, 
        repr.plot.height=6,
        repr.plot.res = 300)

T_vs_N_stems %>%  
tibble::rownames_to_column("gene_symbol") %>%
ggplot(aes(x = avg_log2FC, y = -log(p_val_adj),
               color = I(ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, 'navyblue', 'grey')),
               label = ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, gene_symbol,''))) +
geom_point(size = 1) + 
plot_theme_tmod  + 
geom_text_repel(force = 2, force_pull = 1, 
                box.padding = 1, point.padding = 1,
                max.overlaps = 30, direction = 'both',
                min.segment.length = 0.1,
                aes(color = 'grey10', segment.linetype = 3)) +
labs(title = paste0('Volcano plot of stem cells from\ntumour samples vs normal samples')) +
theme(plot.title = element_text(hjust = 0.5)) +
xlim(-1.5,1.5)
    

### cluster 9 stems vs normal stems

table(subset(adata, Uhlitz_scANVI == 'Stem')@meta.data[c('louvain', 'sample_origin')])

c9_stem = subset(adata, Uhlitz_scANVI == 'Stem' & louvain == '9' & sample_origin == 'tumour\nsample')

normal_stem = subset(adata, Uhlitz_scANVI == 'Stem' & sample_origin == '\nnormal\nsample\n')

c9_normal_stem = merge(c9_stem, normal_stem)

c9_vs_N_stems = FindMarkers(c9_normal_stem,
                            subset.ident = c('Stem'), 
                            group.by = 'louvain', ident.1 = '9')

c9_vs_N_stems

c9_vs_N_stems

c9_vs_N_stems %>%  
tibble::rownames_to_column("gene_symbol") %>%
ggplot(aes(x = avg_log2FC, y = -log(p_val_adj),
               color = I(ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, 'navyblue', 'grey')),
               label = ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, gene_symbol,''))) +
geom_point(size = 1.5) + 
plot_theme_tmod  + 
geom_text_repel(force = 2, force_pull = 1, 
                box.padding = 1, point.padding = 1,
                max.overlaps = 30, direction = 'both',
                min.segment.length = 0.1,
                aes(color = 'grey10', segment.linetype = 3)) +
labs(title = paste0('Volcano plot of cluster 9 stem cells\nvs normal stem cells')) +
theme(plot.title = element_text(hjust = 0.5)) +
xlim(-1.5, 1.5)
    

### tumour sample stem in 9  vs all other stems
for the three cell types: Stem, Stem/TA, Enterocytes 

table(subset(adata, louvain == '9' & sample_origin == 'tumour\nsample')@meta.data$Uhlitz_scANVI)

c9_vs_other_stems = FindMarkers(adata,
                           subset.ident = c('Stem'), 
                           group.by = 'louvain', ident.1 = '9')

c9_vs_other_stems

c9_vs_other_stems %>%  
tibble::rownames_to_column("gene_symbol") %>%
ggplot(aes(x = avg_log2FC, y = -log(p_val_adj),
               color = I(ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, 'navyblue', 'grey')),
               label = ifelse(p_val_adj < 0.05 & abs(avg_log2FC) > 0.5, gene_symbol,''))) +
geom_point(size = 2) + 
plot_theme_tmod  + 
geom_text_repel(force = 2, force_pull = 1, 
                box.padding = 1, point.padding = 1,
                max.overlaps = 30, direction = 'both',
                min.segment.length = 0.1,
                aes(color = 'grey10', segment.linetype = 3)) +
labs(title = paste0('Volcano plot of cluster 9 stem cells\nvs all other stem cells')) +
theme(plot.title = element_text(hjust = 0.5)) +
xlim(-1.5, 1.5)
    

c9_vs_N_stemta = FindMarkers(adata,
                           subset.ident = c('Stem_TA'), 
                           group.by = 'louvain', ident.1 = '9')

c9_vs_N_stemta

c9_vs_N_ep = FindMarkers(adata,
                           subset.ident = c('Enterocyte_progenitor'), 
                           group.by = 'louvain', ident.1 = '9')

c9_vs_N_ep

c9_vs_N_ent = FindMarkers(adata,
                           subset.ident = c('Enterocytes'), 
                           group.by = 'louvain', ident.1 = '9')

c9_vs_N_ent[which(c9_vs_N_ent$p_val_adj < 0.05),]

### transform to SCE
https://hbctraining.github.io/scRNA-seq_online/lessons/pseudobulk_DESeq2_scrnaseq.html

# counts need to be a matrix
counts = adata@assays$raw_counts@counts
metadata = adata@meta.data

sce <- SingleCellExperiment(assays = list(counts = counts), 
                           colData = metadata)

colnames(colData(sce))

### pseudobulk summing counts

groups = c("cell_identity", "sample", "Uhlitz_scANVI")

# more than 30 mins???
# counts need to be a matrix > real fast
summed <- aggregateAcrossCells(sce,
                               id=colData(sce)[,groups])
summed

# add rownames to summed
# collapse 
colnames(summed) = apply(colData(summed)[,groups], 1, function(x)paste(x, collapse = '_'))

### keep only wanted groups: normal cells and genomically normal cells, no TCs

pseudo_bulk = summed[,str_detect(colnames(summed), 'normal')] 
#pseudo_bulk = pseudo_bulk[,!str_detect(colnames(pseudo_bulk), 'TC[1-4]$')] 

head(pseudo_bulk)

### DESeq2

#### need to do this per cell type...........

res_list = list()
res_list_df = list()

resLFC_list = list()
resLFC_df = list()

cell_type_list[-c(7:10)]

for(ctype in cell_type_list[-c(7:10)]){  # not Tuft
    
    print(paste0('start calculating DEG with ', ctype))
    
    slice = pseudo_bulk[,str_detect(colnames(pseudo_bulk), paste0('_',ctype,'$'))] 
    
    # set up DESeq data set
    dds <- DESeqDataSetFromMatrix(countData = counts(slice), # not normalised counts
                              colData = colData(slice),
                              design= ~ cell_identity + patient)
    print(dim(dds))
    
    # set ref level as normal_sample
    dds$cell_identity <- relevel(dds$cell_identity, ref = "normal_sample")
    
    # filter low gene
    # remove gene with less than 10 counts
    #keep <- rowSums(counts(dds)) >= 10
    #dds <- dds[keep,]
    #print(dim(dds))
    
    # genes that only express in some samples
    #keep <- rowSums(counts(dds) >= 10) >= 5
    #dds <- dds[keep,]
    
    # use recommended params: 
    # http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#recommendations-for-single-cell-analysis
    dds_result <- DESeq(dds, test = 'LRT', reduced = ~ patient,
                        useT = TRUE, minmu = 1e-6, minReplicatesForReplace = Inf)
    
    print(resultsNames(dds_result))
    
    res <- results(dds_result, name = 'cell_identity_genomically_normal_vs_normal_sample')
    
    # collect into one list
    res_list[[ctype]] = res
    
    res_list_df[[ctype]] = as.data.frame(res) %>%
    tibble::rownames_to_column("gene_symbol")
    
    # shrinkage result
    resLFC = lfcShrink(dds_result, coef = 'cell_identity_genomically_normal_vs_normal_sample', 
                                type = 'apeglm' )
    
    resLFC_list[[ctype]] = resLFC
    resLFC_df[[ctype]] = as.data.frame(resLFC) %>%
    tibble::rownames_to_column("gene_symbol")
    
}

names(res_list)
# we lost ~10000 genes if removing not so equally expressed genes

plotMA(res_list[['Stem']])

plotMA(resLFC_list[['Stem']])

# instead of plotMA(res)
plot_MA = function(res_name){
    
    p = as.data.frame(res_list[[res_name]]) %>%
    ggplot() + 
    geom_point(aes(x = baseMean, y = log2FoldChange, 
                   color = I(ifelse(padj < 0.05, 'navyblue', 'grey'))), size = 2) +
    scale_x_continuous(trans = log10_trans()) +
    labs(x = 'mean of normalised counts', y = 'log2 fold change') +
    plot_theme_tmod + 
    labs(title = paste0('MA plot of ', res_name)) +
    theme(plot.title = element_text(hjust = 0.5))
    
    print(p)
}

for(i in names(res_list)){
    plot_MA(i)
}

for(i in names(resLFC_list)){
    plot_MA(i)
}

plot_volcano = function(res_name){
    
    # volcano: lo2FC vs -log(FDR)  
    p = as.data.frame(res_list[[res_name]]) %>%
    tibble::rownames_to_column("gene_symbol") %>%
    ggplot(aes(x = log2FoldChange, y = -log(padj),
               color = I(ifelse(padj < 0.05 & abs(log2FoldChange) > 1.5, 'navyblue', 'grey')),
               label = ifelse(padj < .05 & abs(log2FoldChange) > 1.5, gene_symbol,''))) +
    geom_point(size = 2) + 
    plot_theme_tmod  + 
    geom_text_repel(force = 2, force_pull = 1, 
                box.padding = 1, point.padding = 1,
                max.overlaps = 30, direction = 'both',
                min.segment.length = 0.1,
                aes(color = 'grey10', segment.linetype = 3)) +
    labs(title = paste0('Volcano plot of ', res_name, '\n(normal cells from tumour sample vs from normal sample)')) +
    theme(plot.title = element_text(hjust = 0.5))
    
    print(p)
}

for(i in names(res_list)){
    plot_volcano(i)
}

for(i in names(resLFC_list)){
    plot_volcano(i)
}

### dot plot x  = cell type, y = gene, facet by pathway, color by FC, size by padj

for(i in names(res_list_df)){
    res_list_df[[i]]['cell_type'] = i
    res_list_df[[i]] = res_list_df[[i]][c('gene_symbol', 'log2FoldChange', 'padj', 'cell_type')] %>% 
    arrange(padj, abs(log2FoldChange)) #%>%
    #filter(padj < 0.05 & abs(log2FoldChange) >= 1 & gene_symbol %in% R_list) 
}

length(res_list_df)

dim(bind_rows(res_list_df))

res_list_df_full = bind_rows(res_list_df)

res_list_df_full$cell_type = factor(res_list_df_full$cell_type, levels = cell_type_order)

for(i in names(resLFC_df)){
    resLFC_df[[i]]['cell_type'] = i
    resLFC_df[[i]] = resLFC_df[[i]][c('gene_symbol', 'log2FoldChange', 'padj', 'cell_type')] %>% 
    arrange(padj, abs(log2FoldChange)) #%>%
    #filter(padj < 0.05 & abs(log2FoldChange) >= 1 & gene_symbol %in% R_list) 
}

length(resLFC_df)

dim(bind_rows(resLFC_df))

resLFC_df_full = bind_rows(resLFC_df)
resLFC_df_full$cell_type = factor(resLFC_df_full$cell_type, levels = cell_type_order)

### receptor expression 

plot_theme_font_size = function(fontsize = 8){
    plot_theme =  theme(title = element_text(size = fontsize, family = "dajavu sans", colour = 'black', hjust = 0.5),
                   axis.title = element_text(size = fontsize, family = "dajavu sans", colour = 'black'),
                   axis.text = element_text(size = fontsize, family = "dajavu sans", colour = 'black'),
                   legend.title =  element_text(size = fontsize, family = "dajavu sans", colour = 'black'),
                   legend.text=element_text(size=fontsize, family = "dajavu sans", colour = 'black'),
                   text = element_text(size = fontsize, family = "dajavu sans", colour = 'black'),
                   strip.text = element_text(size = fontsize, family = "dajavu sans", colour = 'black'),
                   plot.margin = margin(0,0,0,0,"cm")) 
    plot_theme
}

options(repr.plot.width=8, 
        repr.plot.height=18,
        repr.plot.res = 300)

R_list_exp = as.data.frame(FetchData(adata, vars = R_list)) %>% rownames_to_column()
cell_type_df =  as.data.frame(adata@meta.data[c('Uhlitz_scANVI', 'cell_identity')])  %>% rownames_to_column()

cell_type_df$cell_identity = factor(cell_type_df$cell_identity, 
                                    levels = c('normal_sample', 'genomically_normal', 'genomically_tumour')) 
cell_type_df$Uhlitz_scANVI = factor(cell_type_df$Uhlitz_scANVI, 
                                         levels = cell_type_order)


g = R_list_exp %>% 
left_join(cell_type_df, by = c('rowname' = 'rowname')) %>%
pivot_longer(!c(Uhlitz_scANVI, rowname, cell_identity), names_to = 'gene_symbol', values_to = 'expression') %>%
group_by(Uhlitz_scANVI, gene_symbol, cell_identity) %>%
mutate(mean_expression = mean(expression), ncell = n(), 
       nonzero = ifelse(expression != 0, 1, 0),
       nonzerocell = sum(nonzero),
       cell_fraction = nonzerocell/ncell) %>%
ungroup() %>%
distinct(Uhlitz_scANVI, gene_symbol, cell_identity, mean_expression, cell_fraction) %>%
ggplot(aes(x = Uhlitz_scANVI, y = gene_symbol)) +
geom_tile(fill = 'white') +
#coord_equal() +
geom_point(aes(size = cell_fraction, color = mean_expression)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90),
      strip.background = element_blank(), 
      strip.text = element_text(size = 13), 
      panel.spacing = unit(1, "lines")) +
geom_vline(xintercept = seq(0.5, 9), color = 'grey') +
geom_hline(yintercept = seq(0.5, 54), color = 'grey') +
facet_grid(cols = vars(cell_identity), space = 'free', scale = 'free') +
scale_y_discrete(limits = rev) +
plot_theme_font_size(8)

print(g)

ggsave(file="S6C_R_in_epi_grid.svg", width = 14, height = 30 , units = "cm", dpi = 300)

### differential

options(repr.plot.width=6, 
        repr.plot.height=15,
        repr.plot.res = 300)

res_list_df_full %>%
filter(gene_symbol %in% R_list) %>%
ggplot(aes(x = cell_type, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = -log(padj), color = log2FoldChange)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
#scale_radius(range = c(5, 10)) +
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90)) +
geom_vline(xintercept = seq(0.5, length(res_list_df)), color = 'grey') +
geom_hline(yintercept = seq(0.5, dim(bind_rows(res_list_df))[1]+1), color = 'grey') +
scale_y_discrete(limits = rev)

options(repr.plot.width=6, 
        repr.plot.height=6,
        repr.plot.res = 300)

res_list_df_full %>%
filter(gene_symbol %in% R_list) %>%
filter(padj < 0.05) %>%
ggplot(aes(x = cell_type, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = -log(padj), color = log2FoldChange)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
#scale_radius(range = c(5, 10)) +
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90)) +
geom_vline(xintercept = seq(0.5, length(res_list_df)), color = 'grey') +
geom_hline(yintercept = seq(0.5, dim(bind_rows(res_list_df))[1]+1), color = 'grey') +
scale_y_discrete(limits = rev)

#### with shrinkage

options(repr.plot.width=6, 
        repr.plot.height=15,
        repr.plot.res = 300)

resLFC_df_full %>%
filter(gene_symbol %in% R_list) %>%
ggplot(aes(x = cell_type, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = -log(padj), color = log2FoldChange)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
#scale_radius(range = c(5, 10)) +
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90)) +
geom_vline(xintercept = seq(0.5, length(res_list_df)), color = 'grey') +
geom_hline(yintercept = seq(0.5, dim(bind_rows(res_list_df))[1]+1), color = 'grey') +
scale_y_discrete(limits = rev)

options(repr.plot.width=6, 
        repr.plot.height=6,
        repr.plot.res = 300)

resLFC_df_full %>%
filter(gene_symbol %in% R_list) %>%
mutate(if_plot = ifelse(padj < 0.05, 1, 0)) %>%
group_by(gene_symbol) %>%
mutate(if_plot = sum(if_plot, na.rm = TRUE)) %>%
filter(if_plot >= 1) %>%
ungroup() %>%
ggplot(aes(x = cell_type, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = -log(padj), color = log2FoldChange)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
#scale_radius(range = c(5, 10)) +
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90)) +
geom_vline(xintercept = seq(0.5, length(res_list_df)), color = 'grey') +
geom_hline(yintercept = seq(0.5, dim(bind_rows(res_list_df))[1]+1), color = 'grey') +
scale_y_discrete(limits = rev)

options(repr.plot.width=6, 
        repr.plot.height=6,
        repr.plot.res = 300)

resLFC_df_full %>%
filter(gene_symbol %in% R_list) %>%
filter(padj < 0.05) %>%
ggplot(aes(x = cell_type, y = gene_symbol)) +
geom_tile(fill = 'white') +
coord_equal() +
geom_point(aes(size = -log(padj), color = log2FoldChange)) +
scale_color_gradient2(low = 'blue3', high = 'red3', mid = 'grey85')+
#scale_radius(range = c(5, 10)) +
plot_theme_tmod +
theme(panel.grid = element_blank(),
      axis.text.x = element_text(angle = 90)) +
geom_vline(xintercept = seq(0.5, length(res_list_df)), color = 'grey') +
geom_hline(yintercept = seq(0.5, dim(bind_rows(res_list_df))[1]+1), color = 'grey') +
scale_y_discrete(limits = rev)
